requirements.txt
Flask>=2.2
gunicorn>=20.1
pytest>=7.0
PyJWT>=2.8
pymongo>=4.5
boto3>=1.26
requests>=2.28
flasgger>=0.9.5
python-dotenv>=0.19.0
